feat: add restore end point
