+++ src/api/routes/post.router.ts
@@ -15,8 +15,6 @@ import { getPostUseCaseSchema } from '../useCases/post/getPost.schema.js';
 import { GetPostUseCase } from '../useCases/post/getPost.useCase.js';
 import { getPostsUseCaseSchema } from '../useCases/post/getPosts.schema.js';
 import { GetPostsUseCase } from '../useCases/post/getPosts.useCase.js';
-import { getTrashedPostsUseCaseSchema } from '../useCases/post/getTrashedPosts.schema.js';
-import { GetTrashedPostsUseCase } from '../useCases/post/getTrashedPosts.useCase.js';
 import { trashPostUseCaseSchema } from '../useCases/post/trashPost.schema.js';
 import { TrashPostUseCase } from '../useCases/post/trashPost.useCase.js';
 
@@ -40,29 +38,6 @@ router.get(
     const posts = await useCase.execute(validated.data);
 
     res.json({ posts });
-  })
-);
-
-router.get(
-  '/posts?status=trashed',
-  authenticatedUser,
-  validateAccess(['trashPost']),
-  asyncHandler(async (req: Request, res: Response) => {
-    const validated = getTrashedPostsUseCaseSchema.safeParse({
-      projectId: res.projectRole?.id,
-      defaultLocale: res.projectRole?.defaultLocale,
-    });
-    if (!validated.success) throw new InvalidPayloadException('bad_request', validated.error);
-
-    const useCase = new GetTrashedPostsUseCase(
-      projectPrisma(validated.data.projectId),
-      new PostRepository()
-    );
-    const posts = await useCase.execute(validated.data);
-
-    res.json({
-      posts,
-    });
   })
 );
 
@@ -141,32 +116,6 @@ router.delete(
 
 router.patch(
   '/posts/:id/changeStatus',
-  authenticatedUser,
-  validateAccess(['archivePost', 'publishPost']),
-  asyncHandler(async (req: Request, res: Response) => {
-    const validated = changeStatusUseCaseSchema.safeParse({
-      id: req.params.id,
-      projectId: res.projectRole?.id,
-      userId: res.user.id,
-      status: req.body.status,
-    });
-    if (!validated.success) throw new InvalidPayloadException('bad_request', validated.error);
-
-    const useCase = new ChangeStatusUseCase(
-      projectPrisma(validated.data.projectId),
-      new PostRepository(),
-      new ContentHistoryRepository(),
-      new ContentRepository()
-    );
-
-    await useCase.execute(validated.data);
-
-    res.status(204).send();
-  })
-);
-
-router.patch(
-  '/posts/:id/restore',
   authenticatedUser,
   validateAccess(['archivePost', 'publishPost']),
   asyncHandler(async (req: Request, res: Response) => {
