+++ src/api/useCases/post/restorePost.schema.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+
+export const restorePostUseCaseSchema = z.object({
+  id: z.string(),
+  projectId: z.string(),
+});
+
+export type RestorePostUseCaseSchemaType = z.infer<typeof restorePostUseCaseSchema>;
